refactor: Refactor use of EntityFactory

+++ star-dash/star-dash/GameEngine/Components/PowerUps/SpeedBoostComponent.swift
@@ -1,17 +1,17 @@
 import Foundation
 
 class SpeedBoostComponent: Component {
-    let multiplier: Float
+    let multiplier: CGFloat
     var duration: Float
     var isActivated = false
 
-    init(id: ComponentId, entityId: EntityId, duration: Float, multiplier: Float) {
+    init(id: ComponentId, entityId: EntityId, duration: Float, multiplier: CGFloat) {
         self.duration = duration
         self.multiplier = multiplier
         super.init(id: id, entityId: entityId)
     }
 
-    convenience init(entityId: EntityId, duration: Float, multiplier: Float) {
+    convenience init(entityId: EntityId, duration: Float, multiplier: CGFloat) {
         self.init(id: UUID(), entityId: entityId, duration: duration, multiplier: multiplier)
     }
 }
